Refactor dns endpoints into separate files and common modules
