+++ src/dns/mod.rs
@@ -1,0 +1,21 @@
+/* Pi-hole: A black hole for Internet advertisements
+*  (c) 2018 Pi-hole, LLC (https://pi-hole.net)
+*  Network-wide ad blocking via your own hardware.
+*
+*  API
+*  DNS API Endpoints
+*
+*  This file is copyright under the latest version of the EUPL.
+*  Please see LICENSE file for your rights under this license. */
+
+mod common;
+mod list;
+mod get_list;
+mod add_list;
+mod delete_list;
+mod status;
+
+pub use self::get_list::*;
+pub use self::add_list::*;
+pub use self::delete_list::*;
+pub use self::status::*;
